+++ app/src/main/kotlin/net/ginteam/carmen/kotlin/model/realm/History.kt
@@ -12,7 +12,7 @@ open class HistoryModel : RealmObject() {
     open var id: Long = 0
     open var date: Date? = Date()
     open var price: Double = 0.0
-    open var odometer: Double = 0.0
+    open var odometer: Int = 0
     open var comment: String = ""
 
     open var costType: CostTypeModel? = null
